+++ src/main/kotlin/domain/step2/strategy/calculate/CalculateStrategy.kt
@@ -1,0 +1,6 @@
+package domain.step2.strategy.calculate
+
+@FunctionalInterface
+fun interface CalculateStrategy {
+    fun calculate(x: Int, y: Int): Int
+}
