+++ src/main/kotlin/domain/step2/strategy/split/SplitStrategy.kt
@@ -1,0 +1,6 @@
+package domain.step2.strategy.split
+
+@FunctionalInterface
+fun interface SplitStrategy {
+    fun split(target: String): List<String>
+}
