+++ src/main/kotlin/domain/step2/ui/CalculatorOutputView.kt
@@ -1,0 +1,7 @@
+package domain.step2.ui
+
+import global.strategy.output.OutputStrategy
+
+class CalculatorOutputView(private val outputStrategy: OutputStrategy) {
+    fun outputResult(result: String) = outputStrategy.execute(result)
+}
