+++ src/main/kotlin/global/strategy/input/ConsoleInputStrategy.kt
@@ -1,0 +1,5 @@
+package global.strategy.input
+
+class ConsoleInputStrategy : InputStrategy {
+    override fun execute(): String? = readLine()
+}
