+++ src/main/kotlin/global/strategy/input/InputStrategy.kt
@@ -1,0 +1,6 @@
+package global.strategy.input
+
+@FunctionalInterface
+fun interface InputStrategy {
+    fun execute(): String?
+}
