+++ src/main/kotlin/global/strategy/output/ConsoleOutputStrategy.kt
@@ -1,0 +1,5 @@
+package global.strategy.output
+
+class ConsoleOutputStrategy : OutputStrategy {
+    override fun execute(result: String) = println("연산 결과 : $result")
+}
