+++ src/main/kotlin/global/strategy/output/OutputStrategy.kt
@@ -1,0 +1,6 @@
+package global.strategy.output
+
+@FunctionalInterface
+interface OutputStrategy {
+    fun execute(result: String)
+}
