--- src/main/kotlin/step2/strategy/calculate/CalculateStrategy.kt
@@ -1,6 +1,0 @@
-package step2.strategy.calculate
-
-@FunctionalInterface
-fun interface CalculateStrategy {
-    fun calculate(x: Int, y: Int): Int
-}
