--- src/main/kotlin/step2/strategy/input/InputStrategy.kt
@@ -1,6 +1,0 @@
-package step2.strategy.input
-
-@FunctionalInterface
-fun interface InputStrategy {
-    fun execute(): String?
-}
