--- src/main/kotlin/step2/strategy/split/SingleBlankSplitStrategy.kt
@@ -1,10 +1,0 @@
-package step2.strategy.split
-
-class SingleBlankSplitStrategy : SplitStrategy {
-
-    override fun split(target: String) = target.split(SINGLE_BLANK)
-
-    companion object {
-        private const val SINGLE_BLANK = " "
-    }
-}
