--- src/main/kotlin/step2/strategy/split/SplitStrategy.kt
@@ -1,6 +1,0 @@
-package step2.strategy.split
-
-@FunctionalInterface
-fun interface SplitStrategy {
-    fun split(target: String): List<String>
-}
